Revert "IAM: Add role to User (#111256)"
This reverts commit c012a95fa1592e8257f8ee7ecb44e78b14eb9f80.
