+++ apps/iam/pkg/apis/iam_manifest.go
@@ -11,8 +11,6 @@ import (
 
 	"github.com/grafana/grafana-app-sdk/app"
 	"github.com/grafana/grafana-app-sdk/resource"
-	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/kube-openapi/pkg/spec3"
 
 	v0alpha1 "github.com/grafana/grafana/apps/iam/pkg/apis/iam/v0alpha1"
 )
@@ -95,10 +93,6 @@ var appManifestData = app.ManifestData{
 					Conversion: false,
 				},
 			},
-			Routes: app.ManifestVersionRoutes{
-				Namespaced: map[string]spec3.PathProps{},
-				Cluster:    map[string]spec3.PathProps{},
-			},
 		},
 	},
 }
@@ -136,50 +130,10 @@ var customRouteToGoResponseType = map[string]any{}
 // ManifestCustomRouteResponsesAssociator returns the associated response go type for a given kind, version, custom route path, and method, if one exists.
 // kind may be empty for custom routes which are not kind subroutes. Leading slashes are removed from subroute paths.
 // If there is no association for the provided kind, version, custom route path, and method, exists will return false.
-// Resource routes (those without a kind) should prefix their route with "<namespace>/" if the route is namespaced (otherwise the route is assumed to be cluster-scope)
 func ManifestCustomRouteResponsesAssociator(kind, version, path, verb string) (goType any, exists bool) {
 	if len(path) > 0 && path[0] == '/' {
 		path = path[1:]
 	}
 	goType, exists = customRouteToGoResponseType[fmt.Sprintf("%s|%s|%s|%s", version, kind, path, strings.ToUpper(verb))]
 	return goType, exists
-}
-
-var customRouteToGoParamsType = map[string]runtime.Object{}
-
-func ManifestCustomRouteQueryAssociator(kind, version, path, verb string) (goType runtime.Object, exists bool) {
-	if len(path) > 0 && path[0] == '/' {
-		path = path[1:]
-	}
-	goType, exists = customRouteToGoParamsType[fmt.Sprintf("%s|%s|%s|%s", version, kind, path, strings.ToUpper(verb))]
-	return goType, exists
-}
-
-var customRouteToGoRequestBodyType = map[string]any{}
-
-func ManifestCustomRouteRequestBodyAssociator(kind, version, path, verb string) (goType any, exists bool) {
-	if len(path) > 0 && path[0] == '/' {
-		path = path[1:]
-	}
-	goType, exists = customRouteToGoRequestBodyType[fmt.Sprintf("%s|%s|%s|%s", version, kind, path, strings.ToUpper(verb))]
-	return goType, exists
-}
-
-type GoTypeAssociator struct{}
-
-func NewGoTypeAssociator() *GoTypeAssociator {
-	return &GoTypeAssociator{}
-}
-
-func (g *GoTypeAssociator) KindToGoType(kind, version string) (goType resource.Kind, exists bool) {
-	return ManifestGoTypeAssociator(kind, version)
-}
-func (g *GoTypeAssociator) CustomRouteReturnGoType(kind, version, path, verb string) (goType any, exists bool) {
-	return ManifestCustomRouteResponsesAssociator(kind, version, path, verb)
-}
-func (g *GoTypeAssociator) CustomRouteQueryGoType(kind, version, path, verb string) (goType runtime.Object, exists bool) {
-	return ManifestCustomRouteQueryAssociator(kind, version, path, verb)
-}
-func (g *GoTypeAssociator) CustomRouteRequestBodyGoType(kind, version, path, verb string) (goType any, exists bool) {
-	return ManifestCustomRouteRequestBodyAssociator(kind, version, path, verb)
 }
