--- pkg/registry/apis/iam/common/models.go
@@ -1,8 +1,0 @@
-package common
-
-import "github.com/grafana/grafana/pkg/services/user"
-
-type UserWithRole struct {
-	user.User
-	Role string
-}
