+++ pkg/registry/apis/iam/legacy/user.go
@@ -98,7 +98,7 @@ type ListUserQuery struct {
 }
 
 type ListUserResult struct {
-	Users    []common.UserWithRole
+	Users    []user.User
 	Continue int64
 	RV       int64
 }
@@ -166,10 +166,10 @@ func (s *legacySQLStore) queryUsers(ctx context.Context, sql *legacysql.LegacyDa
 	if err == nil {
 		var lastID int64
 		for rows.Next() {
-			u := common.UserWithRole{}
+			u := user.User{}
 			err = rows.Scan(&u.OrgID, &u.ID, &u.UID, &u.Login, &u.Email, &u.Name,
 				&u.Created, &u.Updated, &u.IsServiceAccount, &u.IsDisabled, &u.IsAdmin, &u.EmailVerified,
-				&u.IsProvisioned, &u.LastSeenAt, &u.Role,
+				&u.IsProvisioned, &u.LastSeenAt,
 			)
 			if err != nil {
 				return res, err
@@ -310,7 +310,7 @@ type CreateUserCommand struct {
 }
 
 type CreateUserResult struct {
-	User common.UserWithRole
+	User user.User
 }
 
 type CreateOrgUserCommand struct {
@@ -398,7 +398,7 @@ func (s *legacySQLStore) CreateUser(ctx context.Context, ns claims.NamespaceInfo
 
 	req := newCreateUser(sql, &cmd)
 
-	var createdUser common.UserWithRole
+	var createdUser user.User
 	err = sql.DB.GetSqlxSession().WithTransaction(ctx, func(st *session.SessionTx) error {
 		userQuery, err := sqltemplate.Execute(sqlCreateUserTemplate, req)
 		if err != nil {
@@ -429,26 +429,23 @@ func (s *legacySQLStore) CreateUser(ctx context.Context, ns claims.NamespaceInfo
 			return fmt.Errorf("failed to create org_user relationship: %w", err)
 		}
 
-		createdUser = common.UserWithRole{
-			User: user.User{
-				ID:               userID,
-				UID:              cmd.UID,
-				Login:            cmd.Login,
-				Email:            cmd.Email,
-				Name:             cmd.Name,
-				OrgID:            cmd.OrgID,
-				IsAdmin:          cmd.IsAdmin,
-				IsDisabled:       cmd.IsDisabled,
-				EmailVerified:    cmd.EmailVerified,
-				IsProvisioned:    cmd.IsProvisioned,
-				Salt:             cmd.Salt,
-				Rands:            cmd.Rands,
-				Created:          cmd.Created.Time,
-				Updated:          cmd.Updated.Time,
-				LastSeenAt:       cmd.LastSeenAt.Time,
-				IsServiceAccount: false,
-			},
-			Role: cmd.Role,
+		createdUser = user.User{
+			ID:               userID,
+			UID:              cmd.UID,
+			Login:            cmd.Login,
+			Email:            cmd.Email,
+			Name:             cmd.Name,
+			OrgID:            cmd.OrgID,
+			IsAdmin:          cmd.IsAdmin,
+			IsDisabled:       cmd.IsDisabled,
+			EmailVerified:    cmd.EmailVerified,
+			IsProvisioned:    cmd.IsProvisioned,
+			Salt:             cmd.Salt,
+			Rands:            cmd.Rands,
+			Created:          cmd.Created.Time,
+			Updated:          cmd.Updated.Time,
+			LastSeenAt:       cmd.LastSeenAt.Time,
+			IsServiceAccount: false,
 		}
 
 		return nil
