+++ pkg/registry/apis/iam/register.go
@@ -2,10 +2,12 @@ package iam
 
 import (
 	"context"
+	"fmt"
 	"maps"
 	"strings"
 
 	"github.com/prometheus/client_golang/prometheus"
+	apierrors "k8s.io/apimachinery/pkg/api/errors"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -323,7 +325,7 @@ func (b *IdentityAccessManagementAPIBuilder) Validate(ctx context.Context, a adm
 	case admission.Create:
 		switch typedObj := a.GetObject().(type) {
 		case *iamv0.User:
-			return user.ValidateOnCreate(ctx, typedObj)
+			return b.validateCreateUser(ctx, a, o)
 		case *iamv0.ServiceAccount:
 			return serviceaccount.ValidateOnCreate(ctx, typedObj)
 		case *iamv0.Team:
@@ -336,6 +338,31 @@ func (b *IdentityAccessManagementAPIBuilder) Validate(ctx context.Context, a adm
 		return nil
 	case admission.Connect:
 		return nil
+	}
+
+	return nil
+}
+
+func (b *IdentityAccessManagementAPIBuilder) validateCreateUser(ctx context.Context, a admission.Attributes, o admission.ObjectInterfaces) error {
+	userObj, ok := a.GetObject().(*iamv0.User)
+	if !ok {
+		return nil
+	}
+
+	requester, err := identity.GetRequester(ctx)
+	if err != nil {
+		return apierrors.NewUnauthorized("no identity found")
+	}
+
+	// Temporary validation that the user is not trying to create a Grafana Admin without being a Grafana Admin.
+	if userObj.Spec.GrafanaAdmin && !requester.GetIsGrafanaAdmin() {
+		return apierrors.NewForbidden(iamv0.UserResourceInfo.GroupResource(),
+			userObj.Name,
+			fmt.Errorf("only grafana admins can create grafana admins"))
+	}
+
+	if userObj.Spec.Login == "" && userObj.Spec.Email == "" {
+		return apierrors.NewBadRequest("user must have either login or email")
 	}
 
 	return nil
