+++ pkg/tests/apis/openapi_snapshots/iam.grafana.app-v0alpha1.json
@@ -4155,8 +4155,7 @@
           "grafanaAdmin",
           "login",
           "name",
-          "provisioned",
-          "role"
+          "provisioned"
         ],
         "properties": {
           "disabled": {
@@ -4184,12 +4183,9 @@
             "default": ""
           },
           "provisioned": {
+            "description": "What to do with salt, rands and password?",
             "type": "boolean",
             "default": false
-          },
-          "role": {
-            "type": "string",
-            "default": ""
           }
         }
       },
@@ -6143,8 +6139,7 @@
           "grafanaAdmin",
           "login",
           "name",
-          "provisioned",
-          "role"
+          "provisioned"
         ],
         "properties": {
           "disabled": {
@@ -6172,12 +6167,9 @@
             "default": ""
           },
           "provisioned": {
+            "description": "What to do with salt, rands and password?",
             "type": "boolean",
             "default": false
-          },
-          "role": {
-            "type": "string",
-            "default": ""
           }
         }
       },
